Rename project reducer
